Added cluster-size condition to Cassandra ctor
An optional configuration setting now forces the embedded Cassandra
storage manager to block until either a timeout elapses or the size of
the Cassandra ring reaches the minimum size specified by the setting.
The timeout is also configurable.  Doing this in the constructor is
probably unwise... it is already pretty elaborate at this point.  We
can refactor later.